# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PlaceMIT - College Placement Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job/internship listings
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [full-time, internship, part-time]
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a specific job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Companies
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # Students
...
  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List placement experiences
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: outcome
          in: query
          schema:
            type: string
            enum: [selected, rejected, on-hold]
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
    post:
      operationId: createExperience
      tags: [experiences]
      summary: Share a placement experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /experiences/{id}:
    get:
      operationId: getExperience
      tags: [experiences]
      summary: Get a placement experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experience details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"

  # Documents
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents in the file explorer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Applications
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List job applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Dashboard / Stats (Wow endpoints)
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, companyId, companyName, type, location, stipend, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        type:
          type: string
          enum: [full-time, internship, part-time]
        location:
          type: string
        stipend:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, companyId, type, location]
      properties:
        title:
          type: string
        companyId:
          type: integer
        type:
          type: string
          enum: [full-time, internship, part-time]
        location:
          type: string
        stipend:
          type: string
        description:
          type: string
        requirements:
          type: string
        deadline:
          type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        location:
          type: string
        stipend:
          type: string
        description:
          type: string
        deadline:
          type: string

    # Company schemas
    Company:
      type: object
      required: [id, name, sector, logoUrl, placementsCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        placementsCount:
          type: integer
        avgPackage:
          type: ["string", "null"]
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name, sector]
      properties:
        name:
          type: string
        sector:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        description:
          type: string
        avgPackage:
          type: string

    # Student schemas
...
    Experience:
      type: object
      required: [id, studentName, companyName, role, outcome, rounds, createdAt]
      properties:
        id:
          type: integer
        studentName:
          type: string
        studentAvatarUrl:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: string
        role:
          type: string
        outcome:
          type: string
          enum: [selected, rejected, on-hold]
        rounds:
          type: integer
        description:
          type: ["string", "null"]
        tips:
          type: ["string", "null"]
        packageOffered:
          type: ["string", "null"]
        createdAt:
          type: string
    ExperienceInput:
      type: object
      required: [studentName, companyName, role, outcome, rounds]
      properties:
        studentName:
          type: string
        studentAvatarUrl:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        role:
          type: string
        outcome:
          type: string
          enum: [selected, rejected, on-hold]
        rounds:
          type: integer
        description:
          type: string
        tips:
          type: string
        packageOffered:
          type: string

    # Document schemas
    Document:
      type: object
      required: [id, name, type, url, studentId, studentName, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [resume, offer-letter, certificate, other]
        url:
          type: string
        studentId:
          type: integer
        studentName:
          type: string
        sizeKb:
          type: ["integer", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [name, type, url, studentId, studentName]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [resume, offer-letter, certificate, other]
        url:
          type: string
        studentId:
          type: integer
        studentName:
          type: string
        sizeKb:
          type: integer

    # Application schemas
    Application:
      type: object
      required: [id, studentId, studentName, jobId, jobTitle, companyName, status, appliedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        jobId:
          type: integer
        jobTitle:
          type: string
        companyName:
          type: string
        status:
          type: string
          enum: [applied, shortlisted, interview, selected, rejected]
        appliedAt:
          type: string
    ApplicationInput:
      type: object
      required: [studentId, studentName, jobId, jobTitle, companyName]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        jobId:
          type: integer
        jobTitle:
          type: string
        companyName:
          type: string
    ApplicationUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [applied, shortlisted, interview, selected, rejected]

    # Dashboard / Stats schemas
...