# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlaceMIT - College Placement Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  sector: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  description?: string | null;
  placementsCount: number;
  /** @nullable */
  avgPackage?: string | null;
  createdAt: string;
}