# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlaceMIT - College Placement Community Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentType } from './documentType';

export interface Document {
  id: number;
  name: string;
  type: DocumentType;
  url: string;
  studentId: number;
  studentName: string;
  /** @nullable */
  sizeKb?: number | null;
  createdAt: string;
}