# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlaceMIT - College Placement Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  title?: string;
  type?: string;
  location?: string;
  stipend?: string;
  description?: string;
  deadline?: string;
}